IT_forum/requirement.txt
alembic==1.0.0
aniso8601==3.0.2
asn1crypto==0.24.0
blinker==1.4
cffi==1.11.5
click==6.7
cryptography==2.3
Django==1.11.4
dominate==2.3.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Cache==0.13.1
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Moment==0.6.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
Pillow==5.2.0
pycparser==2.18
PyMySQL==0.9.2
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
six==1.11.0
SQLAlchemy==1.2.10
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
